rgw: false assumption on vault bucket key deletion
On bucket key deletion when the request to change the property of
the key for deletion_allowed to true, it is expected that the
response body be empty. But this assumption is false and there
would be a dump of the new config in the response. this condition
would prevent the key deletion to be done.

Resolves: rhbz#2279912
Fixes: https://tracker.ceph.com/issues/65626
Signed-off-by: Seena Fallah <seenafallah@gmail.com>
(cherry picked from commit d76e1be535976597c9b57d3b0153cc4630a64126)