mgr/object_format: allow easily disabling dict key sorting
When adapating an object to JSON or YAML formatting the previous version
of the code always sorted dict keys. This is nice for code that placed
no ordering on the dict keys. However, for newer code that may rely on
python's sorted dicts we want to be able to disable this behavior.
Add two new keyword args to the adapter type that allow disabling
the sorting.

Resolves: rhbz#2300005
Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit 53e81500e8cb07f4788a6eba8ed1cc5dd8421423)