mgr/orchestrator: add default for group field in nvmeof spec
We use the group attribute for the `ceph nvme-gw
create` and `ceph nvme-gw delete` commands that
are run while deploying and removing nvmeof
gateways. Therefore we need this field to be
set. This patch for now defaults it to "default".
We may later require users to set this field
when creating new nvmeof services.

Resolves: rhbz#2282364

Signed-off-by: Adam King <adking@redhat.com>