mgr/smb: accept public_addrs on cli when creating cluster
We can set the public ip address to set for the cluster using the
declarative method by providing the information in the resource
description. The corresponding functionality is not available with the
imperative method of creating the smb cluster.

This patch adds this functionality by allowing the user the option of
providing the a public address on the command line when creating the
smb cluster.

Signed-off-by: Sachin Prabhu <sp@spui.uk>
(cherry picked from commit 4db3bb77b6458f8b54af7f9279151d616f042d49)

Resolves: rhbz#2311116

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>