mgr/cephadm: fix custom alertmanager webhooks
Previously, we put the custom webhooks in the "default"
receiver. This didn't actually work as alertmanager only sends
alerts to the default receiver if the alert does not match
the following routes. This meant if you wanted alerts sent
to the dashboard and also a custom location, you weren't
able to do so with the template cephadm provided.

This also swaps to using "webhook_urls" instead of
"default_webhook_urls" for what should be in the spec,
but "default_webhook_urls" is kept working for backwards
compatability

Fixes: https://tracker.ceph.com/issues/68157

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 7482bb7bf65cc5e0dd681c96b9c88db49010557d)

Resolves: rhbz#2313614