mgr/callhome: persist operations between mgr restarts
Currently the operations dictionary is only kept in memory. It is lost
when the mgr restarts, and this can cause the module to handle upload
requests which were already processed and registered in the operations
dictionary. To prevent that, we write the operations to the db, and load
them when the module starts.

Resolves: rhbz#2317147

https://bugzilla.redhat.com/show_bug.cgi?id=2317147

Signed-off-by: Yaarit Hatuka <yhatuka@ibm.com>
(cherry picked from commit 9a28b7c97467ede85afdb8d71a19b0d7be124280)