rgw/cloudtier: Correct option ordering in RGWZoneGroupPlacementTier
Two tier-config options (related to `cloud-restore`) were incorrectly added
in the middle of the encoding and decoding methods of RGWZoneGroupPlacementTier.
This modification can cause compatibility issues with older decoders when
attempting to read v2-encoded REST objects.

The fix is to correct the option order and update the decode() function to
properly interpret the structure based on the encoded version.

Signed-off-by: Soumya Koduri <skoduri@redhat.com>
Resolves: rhbz#2342847