cephadm/smb: convert config object to use a dataclass
While working on the smb.py file I began to be annoyed at the Config
class and the need to repeat myself. Now that cephadm is not expected to
run on python versions older than 3.7 I think it's safe to convert
Config to use a dataclass.

While making the change to a dataclass I also chose to make it a frozen
dataclass to help eliminate any future bugs wrt mutating the config object.

Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit 2b30c8e8958577bb756aefd10bae411138f0ccd5)

Resolves: rhbz#2346093

Fix some parts of the code which were not consistent with upstream
because of the order in which patches were backported to downstream.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>